Reorganized project structure for better modularity and maintainability
- Moved actor-related classes to com.example.demo.actors
- Moved level-related classes to com.example.demo.levels
- Moved projectile-related classes to com.example.demo.projectiles
- Moved UI components to com.example.demo.ui
- Moved utility classes to com.example.demo.utilities
- Ensured imports and references were updated accordingly
- Verified the application runs as expected after reorganization
