+++ src/main/java/com/example/demo/actors/UserPlane.java
@@ -1,4 +1,6 @@
-package com.example.demo;
+package com.example.demo.actors;
+
+import com.example.demo.projectiles.UserProjectile;
 
 import javafx.geometry.BoundingBox;
 import javafx.geometry.Bounds;
