+++ src/main/java/com/example/demo/levels/LevelBoss.java
@@ -1,4 +1,6 @@
-package com.example.demo;
+package com.example.demo.levels;
+
+import com.example.demo.actors.Boss;
 
 import javafx.stage.Stage;
