+++ src/main/java/com/example/demo/levels/LevelOne.java
@@ -1,16 +1,21 @@
-package com.example.demo;
+package com.example.demo.levels;
+
+import com.example.demo.actors.ActiveActorDestructible;
+import com.example.demo.actors.EnemyPlane;
 
 import javafx.stage.Stage;
 
-public class LevelTwo extends LevelParent {
+public class LevelOne extends LevelParent {
 
-    private static final String BACKGROUND_IMAGE_NAME = "/com/example/demo/images/space2.jpg";
-    private static final int TOTAL_ENEMIES = 7; // Number of enemies to spawn
-    private static final int KILLS_TO_ADVANCE = 15; // Kill count required to move to the next level
-    private static final double ENEMY_SPAWN_PROBABILITY = 0.25; // Probability for spawning enemies
+    private static final String BACKGROUND_IMAGE_NAME = "/com/example/demo/images/space.jpg";
+    private static final String NEXT_LEVEL = "com.example.demo.levels.LevelTwo";
+    private static final int TOTAL_ENEMIES = 5;
+    private static final int KILLS_TO_ADVANCE = 10;
+    private static final double ENEMY_SPAWN_PROBABILITY = 0.20;
+    private static final int PLAYER_INITIAL_HEALTH = 5;
 
-    public LevelTwo(double screenHeight, double screenWidth, Stage stage) {
-        super(BACKGROUND_IMAGE_NAME, screenHeight, screenWidth, 5, stage); // Initialize LevelParent with 5 player health
+    public LevelOne(double screenHeight, double screenWidth, Stage stage) {
+        super(BACKGROUND_IMAGE_NAME, screenHeight, screenWidth, PLAYER_INITIAL_HEALTH, stage);
     }
 
     @Override
@@ -21,7 +26,7 @@ public class LevelTwo extends LevelParent {
         }
         // Check if the player has reached the kill target to transition to the next level
         else if (userHasReachedKillTarget()) {
-            goToNextLevel("com.example.demo.LevelBoss"); // Transition to LevelBoss
+            goToNextLevel(NEXT_LEVEL);
         }
         System.out.println("Kill Target Reached: " + getUser().getNumberOfKills());
     }
@@ -47,11 +52,11 @@ public class LevelTwo extends LevelParent {
     @Override
     protected LevelView instantiateLevelView() {
         // Create a LevelView specific to this level
-        return new LevelView(getRoot(), 5); // Display 5 hearts (player health)
+        return new LevelView(getRoot(), PLAYER_INITIAL_HEALTH);
     }
 
-    // Helper method to check if the user has reached the kill target
     private boolean userHasReachedKillTarget() {
+        // Check if the user's kill count has reached the target to advance to the next level
         return getUser().getNumberOfKills() >= KILLS_TO_ADVANCE;
     }
 }
