+++ src/main/java/com/example/demo/projectiles/EnemyProjectile.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.demo.projectiles;
 
 public class EnemyProjectile extends Projectile {
