+++ src/main/java/com/example/demo/ui/HeartDisplay.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.demo.ui;
 
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
