+++ src/main/java/com/example/demo/ui/MainMenu.java
@@ -1,4 +1,7 @@
-package com.example.demo.controller;
+package com.example.demo.ui;
+
+import com.example.demo.controller.Controller;
+import com.example.demo.controller.HowToPlayWindow;
 
 import javafx.application.Application;
 import javafx.geometry.Pos;
